requirements.txt
django
pymongo
numpy
matplotlib
tqdm
pydrive
torch
torchvision
gmplot
pymongo
eventlet
python-socketio
